add mod support

+++ modules/carafe.nf
@@ -13,6 +13,9 @@ process CARAFE {
         path fasta_file
         path peptide_results_file
         val carafe_params
+        val include_phosphorylation
+        val include_oxidized_methionine
+        val max_mod_option
         val output_format
     
     output:
@@ -26,22 +29,34 @@ process CARAFE {
     script:
 
         apptainer_cmds = ''
-        if (workflow.containerEngine == 'singularity' || workflow.containerEngine == 'apptainer') {
-            // Running with Apptainer/Singularity
-            apptainer_cmds = """
-                source /opt/conda/etc/profile.d/conda.sh
-                conda activate carafe
-            """
-        }
+        // if (workflow.containerEngine == 'singularity' || workflow.containerEngine == 'apptainer') {
+        //     // Running with Apptainer/Singularity
+        //     apptainer_cmds = """
+        //         source /opt/conda/etc/profile.d/conda.sh
+        //         conda activate carafe
+        //     """
+        // }
 
         lf_type_param = output_format == 'diann' ? 'diann' : 'encyclopedia'
+
+        // Build variable modification parameters
+        mod_param = ''
+        if(include_phosphorylation && include_oxidized_methionine) {
+            mod_param = "-varMod 2,7,8,9 -mode phosphorylation ${max_mod_option}"
+        } else if(include_phosphorylation) {
+            mod_param = "-varMod 7,8,9 -mode phosphorylation ${max_mod_option}"
+        } else if(include_oxidized_methionine) {
+            mod_param = "-varMod 2 -mode general ${max_mod_option}"
+        } else {
+            mod_param = '-mode general'
+        }
 
         def container_image = task.container ?: 'none'
 
         """
         ${apptainer_cmds}
 
-        export HOME=\$PWD
+        # export HOME=\$PWD
 
         ${exec_java_command(task.memory)} \\
         -ms "." \\
@@ -50,10 +65,12 @@ process CARAFE {
         -se "DIA-NN" \\
         -lf_type ${lf_type_param} \\
         -device cpu \\
+        ${mod_param} \\
         ${carafe_params} \\
         > >(tee "carafe.stdout") 2> >(tee "carafe.stderr" >&2)
 
-        mv -v SkylineAI_spectral_library.tsv carafe_spectral_library.tsv
+        # move SkylineAI_spectral_library.tsv if it exists (legacy compatibility)
+        [ -e SkylineAI_spectral_library.tsv ] && mv -v SkylineAI_spectral_library.tsv carafe_spectral_library.tsv
 
         CARAFE_VERSION=\$(grep "Version:" carafe.stdout | head -n 1 | awk '{print \$2}' || true)
         CARAFE_VERSION=\${CARAFE_VERSION:-unknown}
